Исправлено приложение ktor, чтобы тесты могли запускаться
